requirements.txt
crewai>=0.30.0
tavily-python
yfinance
numpy
pandas
langfuse
openinference-instrumentation-crewai
openinference-instrumentation-litellm
litellm
python-dotenv
groq
fastapi
uvicorn
apscheduler
fastapi_sso
